+++ spec/adivina_la_palabra_spec.rb
@@ -1,4 +1,4 @@
-require "./adivina-la-palabra.rb"
+require "./lib/adivina-la-palabra.rb"
 describe AdivinaLaPalabra do
     # Palabra secreta en astiriscos
     it "Convertir la palabra en asteriscos" do
@@ -10,5 +10,10 @@ describe AdivinaLaPalabra do
         adivinaLaPalabra = AdivinaLaPalabra.new
         asteriscos = adivinaLaPalabra.ocultar_palabra "Inglaterra"
         expect(asteriscos).not_to eq "*****"
+    end
+    it "la letra está en la palabra" do
+        adivinaLaPalabra =  AdivinaLaPalabra.new
+        estaLaPalabra =  adivinaLaPalabra.esta_la_letra? "a"
+        expect(estaLaPalabra).to eq true
     end
 end
